pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "agentprobe"
version = "0.1.0"
description = "Functional agent evaluation CLI for HTTP APIs and local agent harnesses."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "click>=8.1.7",
  "cryptography>=45.0.0",
  "httpx>=0.27.2",
  "jinja2>=3.1.4",
  "jsonpath-ng>=1.7.0",
  "openai>=1.54.0",
  "pydantic>=2.9.2",
  "pyjwt>=2.9.0",
  "pyyaml>=6.0.2",
  "sqlalchemy>=2.0.36",
  "websockets>=15.0.1",
]

[project.scripts]
agentprobe = "agentprobe.cli:cli"

[dependency-groups]
dev = [
  "mypy>=1.18.2",
  "pyright>=1.1.403",
  "pytest>=8.3.3",
  "types-PyYAML>=6.0.12.20250915",
]

[tool.mypy]
python_version = "3.11"
files = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
